@@ -17,76 +17,78 @@
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config ./test/jest-e2e.json",
+    "test:e2e": "jest --config ./test/jest-e2e.json --runInBand",
+    "test:e2e:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/jest/bin/jest.js --config ./test/jest-e2e.json --runInBand",
     "typeorm": "ts-node ./node_modules/typeorm/cli",
     "typeorm:run-migrations": "npm run typeorm migration:run -- -d ./src/typeorm-cli.config.ts",
     "typeorm:generate-migration": "npm run typeorm -- -d ./src/typeorm-cli.config.ts migration:generate ./migrations/migration",
     "typeorm:create-migration": "npm run typeorm -- migration:create ./migrations/",
     "typeorm:revert-migration": "npm run typeorm -- -d ./src/typeorm-cli.config.ts migration:revert",
+    "resetdb": "npm run typeorm schema:drop -- -d ./src/typeorm-cli.config.ts && npm run typeorm:run-migrations",
     "fixtures": "fixtures-ts-node-commonjs"
   },
   "dependencies": {
-    "@nestjs/axios": "^3.0.0",
-    "@nestjs/common": "^10.2.5",
-    "@nestjs/config": "^3.1.1",
-    "@nestjs/core": "^10.2.5",
-    "@nestjs/event-emitter": "^2.0.2",
-    "@nestjs/jwt": "^10.1.1",
-    "@nestjs/passport": "^10.0.2",
-    "@nestjs/platform-express": "^10.2.5",
-    "@nestjs/platform-ws": "^10.2.5",
-    "@nestjs/schedule": "^3.0.3",
-    "@nestjs/swagger": "^7.3.0",
-    "@nestjs/typeorm": "^10.0.0",
-    "@nestjs/websockets": "^10.2.5",
-    "@twurple/auth": "^7.0.2",
-    "@twurple/chat": "^7.0.2",
+    "@nestjs/axios": "^3.0.2",
+    "@nestjs/common": "^10.3.7",
+    "@nestjs/config": "^3.2.1",
+    "@nestjs/core": "^10.3.7",
+    "@nestjs/event-emitter": "^2.0.4",
+    "@nestjs/jwt": "^10.2.0",
+    "@nestjs/passport": "^10.0.3",
+    "@nestjs/platform-express": "^10.3.7",
+    "@nestjs/platform-ws": "^10.3.7",
+    "@nestjs/schedule": "^4.0.1",
+    "@nestjs/swagger": "^7.3.1",
+    "@nestjs/typeorm": "^10.0.2",
+    "@nestjs/websockets": "^10.3.7",
+    "@twurple/auth": "^7.1.0",
+    "@twurple/chat": "^7.1.0",
     "@willsoto/nestjs-prometheus": "^6.0.0",
-    "axios": "^1.5.0",
+    "axios": "^1.6.8",
     "child-process-promise": "^2.2.1",
     "cookie-parser": "^1.4.6",
     "hbs": "^4.2.0",
     "memory-streams": "^0.1.3",
     "neat-csv": "^7.0.0",
-    "nestjs-i18n": "^10.3.2",
-    "passport": "^0.6.0",
+    "nestjs-i18n": "^10.4.5",
+    "passport": "^0.7.0",
     "passport-jwt": "^4.0.1",
     "passport-steam": "^1.0.18",
-    "pg": "^8.11.3",
-    "prom-client": "^15.1.0",
-    "reflect-metadata": "^0.1.13",
+    "pg": "^8.11.4",
+    "prom-client": "^15.1.1",
+    "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1",
-    "sqlite3": "^5.1.6",
-    "typeorm": "^0.3.17",
+    "sqlite3": "^5.1.7",
+    "typeorm": "^0.3.20",
     "uuid": "^9.0.1",
-    "yauzl": "^2.10.0"
+    "yauzl": "^3.1.2"
   },
   "devDependencies": {
-    "@nestjs/cli": "^10.1.17",
-    "@nestjs/schematics": "^10.0.2",
-    "@nestjs/testing": "^10.2.5",
-    "@types/cookie-parser": "^1.4.6",
-    "@types/express": "^4.17.17",
-    "@types/jest": "29.5.5",
-    "@types/node": "20.6.1",
-    "@types/passport-steam": "^1.0.2",
-    "@types/supertest": "^2.0.12",
+    "@nestjs/cli": "^10.3.2",
+    "@nestjs/schematics": "^10.1.1",
+    "@nestjs/testing": "^10.3.7",
+    "@types/cookie-parser": "^1.4.7",
+    "@types/express": "^4.17.21",
+    "@types/jest": "29.5.12",
+    "@types/node": "20.12.2",
+    "@types/passport-steam": "^1.0.5",
+    "@types/supertest": "^6.0.2",
     "@types/yauzl": "^2.10.3",
-    "@typescript-eslint/eslint-plugin": "^6.7.0",
-    "@typescript-eslint/parser": "^6.7.0",
-    "eslint": "^8.49.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-prettier": "^5.0.0",
+    "@typescript-eslint/eslint-plugin": "^7.5.0",
+    "@typescript-eslint/parser": "^7.5.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "jest": "29.7.0",
-    "prettier": "^3.0.3",
+    "prettier": "^3.2.5",
     "source-map-support": "^0.5.21",
-    "supertest": "^6.3.3",
-    "ts-jest": "29.1.1",
-    "ts-loader": "^9.4.4",
-    "ts-node": "^10.9.1",
+    "supertest": "^6.3.4",
+    "ts-jest": "29.1.2",
+    "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.2",
     "tsconfig-paths": "4.2.0",
     "typeorm-fixtures-cli": "^4.0.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.4.3"
   },
   "jest": {
     "moduleFileExtensions": [
