@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { ChangeDetectorRef, Component, Input, OnInit } from '@angular/core';
 import { QueuebotApiService } from '../../services/queuebot-api.service';
 import { HttpClientModule } from '@angular/common/http';
 import { CommonModule } from '@angular/common';
@@ -11,13 +11,43 @@ import {
 import { WebsocketService } from '../../services/websocket.service';
 import { MatIcon } from '@angular/material/icon';
 import { WindowWithElectron } from '../../models/window.global';
+import {
+  MatCell,
+  MatCellDef,
+  MatColumnDef,
+  MatHeaderCell,
+  MatHeaderCellDef,
+  MatHeaderRow,
+  MatHeaderRowDef,
+  MatRow,
+  MatRowDef,
+  MatTable,
+} from '@angular/material/table';
+import { LocalSongStatusComponent } from '../local-song-status/local-song-status.component';
+import { LocalSongState } from '../../models/local-song-state';
 
 declare let window: WindowWithElectron;
 
 @Component({
   selector: 'app-queue-list',
   standalone: true,
-  imports: [HttpClientModule, CommonModule, DragDropModule, MatIcon],
+  imports: [
+    HttpClientModule,
+    CommonModule,
+    DragDropModule,
+    MatIcon,
+    MatTable,
+    MatHeaderCellDef,
+    MatCellDef,
+    MatHeaderCell,
+    MatCell,
+    MatHeaderRow,
+    MatRow,
+    MatHeaderRowDef,
+    MatRowDef,
+    MatColumnDef,
+    LocalSongStatusComponent,
+  ],
   providers: [],
   templateUrl: './queue-list.component.html',
   styleUrl: './queue-list.component.css',
@@ -27,9 +57,17 @@ export class QueueListComponent implements OnInit {
   channelName: string = '';
 
   songRequests: SongRequestDto[] = [];
+
+  downloadedSongStatus: Map<number, LocalSongState> = new Map<
+    number,
+    LocalSongState
+  >();
+
+  columnsToDisplay = ['song', 'songLength', 'requester', 'ops'];
   constructor(
     private queuebotApiService: QueuebotApiService,
     private websocketService: WebsocketService,
+    private ref: ChangeDetectorRef,
   ) {}
 
   ngOnInit() {
@@ -62,7 +100,7 @@ export class QueueListComponent implements OnInit {
         });
       }
 
-      this.websocketService.messages$.subscribe((message) => {
+      this.websocketService.messages$.subscribe(async (message) => {
         console.log(message);
 
         if (message.event == 'songRequestQueueChanged') {
@@ -70,7 +108,8 @@ export class QueueListComponent implements OnInit {
 
           if (window.songs) {
             for (const songRequest of this.songRequests) {
-              window.songs.processSong(songRequest);
+              console.log('Calling processSong', { songRequest: songRequest });
+              await window.songs.processSong(songRequest);
             }
           } else {
             console.log(
@@ -78,6 +117,16 @@ export class QueueListComponent implements OnInit {
             );
           }
         }
+      });
+    }
+
+    // Attach to listening to the download progress thing.
+    if (window.songs) {
+      console.log('Connecting to onProcessSongProgress');
+      window.songs.onProcessSongProgress((songState: LocalSongState) => {
+        console.log('Download update', { songState: songState });
+        this.downloadedSongStatus.set(songState.songId, songState);
+        this.ref.detectChanges();
       });
     }
   }
@@ -119,5 +168,11 @@ export class QueueListComponent implements OnInit {
           console.log('Got error', { err: err });
         },
       });
+  }
+
+  setSongRequestActive(songRequestId: number) {
+    this.queuebotApiService
+      .setSongRequestActive(this.channelName, songRequestId)
+      .subscribe();
   }
 }
