@@ -1,19 +1,23 @@
 import { DownloadHandler } from './handlers/download-handler.interface';
 import { SongDto } from '../../../common';
+import { DownloadState } from '../local-song-state';
 
 export class SongDownloader {
   constructor(private downloadHandlers: DownloadHandler[]) {}
 
-  async processSong(song: SongDto) {
+  async processSong(song: SongDto, songStateCallback: any) {
     for (const downloadHandler of this.downloadHandlers) {
       if (downloadHandler.songSupported(song)) {
         if (downloadHandler.songIsLocal(song)) {
           // Song is already present, no need to download it.
           console.log('Song present', { songId: song.id, title: song.title });
+          songStateCallback({
+            songId: song.id,
+            downloadState: DownloadState.Complete,
+          });
           return;
         }
-        console.log('Downloading song', { songId: song.id, title: song.title });
-        await downloadHandler.downloadSong(song);
+        await downloadHandler.downloadSong(song, songStateCallback);
       }
     }
   }
