@@ -20,9 +20,12 @@ import { environment } from './environment';
 export const IPC_OPEN_TWITCH_LOGIN = 'login.openTwitchLogin';
 export const IPC_SETTINGS_GET_VALUE = 'settings.getValue';
 export const IPC_SETTINGS_SET_VALUE = 'settings.setValue';
-// FIXME: Need to make this configurable at build time
+export const IPC_SETTINGS_DELETE_VALUE = 'settings.deleteValue';
 export const LOGIN_URL = `${environment.queuebotApiBaseUrl}/auth/twitch`;
 export const IPC_SONG_DOWNLOADER_PROCESS_SONG = 'songDownloader.processSong';
+export const IPC_SONG_DOWNLOADER_PROCESS_SONG_PROGRESS =
+  'songDownloader.processSongProgress';
+
 export const IPC_PROTOCOL_HANDLER = 'login.protocolHandler';
 
 export const FILTERED_URLS = [`${environment.queuebotApiBaseUrl}/*`];
@@ -78,6 +81,10 @@ function bootstrap() {
     return settingsService.getValue(key);
   });
 
+  ipcMain.handle(IPC_SETTINGS_DELETE_VALUE, (event, key) => {
+    return settingsService.deleteValue(key);
+  });
+
   ipcMain.handle(IPC_OPEN_TWITCH_LOGIN, (event, args) => {
     console.log('Opening external');
     shell.openExternal(LOGIN_URL).then(() => {
@@ -88,7 +95,12 @@ function bootstrap() {
   ipcMain.handle(
     IPC_SONG_DOWNLOADER_PROCESS_SONG,
     async (event, song: SongDto) => {
-      return await songDownloader.processSong(song);
+      await songDownloader.processSong(song, (songState: any) => {
+        win?.webContents.send(
+          IPC_SONG_DOWNLOADER_PROCESS_SONG_PROGRESS,
+          songState,
+        );
+      });
     },
   );
 
