@@ -11,6 +11,10 @@ const IPC_OPEN_TWITCH_LOGIN = 'login.openTwitchLogin';
 const IPC_SETTINGS_GET_VALUE = 'settings.getValue';
 const IPC_SETTINGS_SET_VALUE = 'settings.setValue';
 const IPC_SONG_DOWNLOADER_PROCESS_SONG = 'songDownloader.processSong';
+const IPC_SETTINGS_DELETE_VALUE = 'settings.deleteValue';
+const IPC_SONG_DOWNLOADER_PROCESS_SONG_PROGRESS =
+  'songDownloader.processSongProgress';
+
 export const IPC_PROTOCOL_HANDLER = 'login.protocolHandler';
 
 contextBridge.exposeInMainWorld('settings', {
@@ -18,6 +22,8 @@ contextBridge.exposeInMainWorld('settings', {
     ipcRenderer.invoke(IPC_SETTINGS_SET_VALUE, key, value),
   getValue: (key: string) => ipcRenderer.invoke(IPC_SETTINGS_GET_VALUE, key),
   openTwitchLogin: () => ipcRenderer.invoke(IPC_OPEN_TWITCH_LOGIN),
+  deleteValue: (key: string) =>
+    ipcRenderer.invoke(IPC_SETTINGS_DELETE_VALUE, key),
 });
 
 // For some reason, this errors out with "
@@ -32,6 +38,11 @@ contextBridge.exposeInMainWorld('login', {
 contextBridge.exposeInMainWorld('songs', {
   processSong: (songDto: SongDto) =>
     ipcRenderer.invoke(IPC_SONG_DOWNLOADER_PROCESS_SONG, songDto),
+  onProcessSongProgress: (callback: any) =>
+    ipcRenderer.on(
+      IPC_SONG_DOWNLOADER_PROCESS_SONG_PROGRESS,
+      (_event, songState) => {
+        callback(songState);
+      },
+    ),
 });
-
-//
