@@ -4,17 +4,11 @@ export class SettingsStoreService {
   constructor(private filePath: string) {}
 
   setValue(key: string, value: string) {
-    let settings: any;
-    if (!fs.existsSync(this.filePath)) {
-      // assume we need to create it.
-      settings = {};
-    } else {
-      settings = JSON.parse(fs.readFileSync(this.filePath).toString());
-    }
+    const settings = this.loadSettings();
 
     settings[key] = value;
 
-    fs.writeFileSync(this.filePath, JSON.stringify(settings));
+    this.saveSettings(settings);
 
     console.log('Wrote setting to settings file', {
       filePath: this.filePath,
@@ -24,7 +18,25 @@ export class SettingsStoreService {
   }
 
   getValue(key: string): string | undefined {
-    let settings: any;
+    const settings = this.loadSettings();
+    if (settings[key]) {
+      return settings[key];
+    }
+    return undefined;
+  }
+
+  deleteValue(key: string): void {
+    const settings = this.loadSettings();
+    if (settings[key]) {
+      delete settings[key];
+    }
+
+    this.saveSettings(settings);
+  }
+
+  private loadSettings(): any {
+    let settings;
+
     if (!fs.existsSync(this.filePath)) {
       // assume we need to create it.
       settings = {};
@@ -32,9 +44,10 @@ export class SettingsStoreService {
       settings = JSON.parse(fs.readFileSync(this.filePath).toString());
     }
 
-    if (settings[key]) {
-      return settings[key];
-    }
-    return undefined;
+    return settings;
+  }
+
+  private saveSettings(settings: any): void {
+    fs.writeFileSync(this.filePath, JSON.stringify(settings));
   }
 }
