@@ -2,7 +2,7 @@ import { Inject, Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable, retry, timer } from 'rxjs';
 import { QUEUEBOT_API_BASE_URL } from '../app.config';
-import { SongRequestDto } from '../../../../common';
+import { GameDto, SongRequestDto } from '../../../../common';
 import { ChannelDto } from '../../../../common';
 
 @Injectable({
@@ -61,6 +61,72 @@ export class QueuebotApiService {
   ): Observable<boolean> {
     return this.httpClient.delete<boolean>(
       `${this.apiBaseUrl}/api/channels/${channel}/song-requests/${songRequestId}`,
+      {
+        withCredentials: true,
+      },
+    );
+  }
+
+  nextSong(channel: string): Observable<SongRequestDto | undefined> {
+    return this.httpClient.put<SongRequestDto>(
+      `${this.apiBaseUrl}/api/channels/${channel}/song-requests/next-song`,
+      {},
+      {
+        withCredentials: true,
+      },
+    );
+  }
+
+  setSongRequestActive(
+    channel: string,
+    songRequestId: number,
+  ): Observable<SongRequestDto> {
+    return this.httpClient.put<SongRequestDto>(
+      `${this.apiBaseUrl}/api/channels/${channel}/song-requests/${songRequestId}`,
+      {
+        songRequestId: songRequestId,
+        isActive: true,
+      },
+      {
+        withCredentials: true,
+      },
+    );
+  }
+
+  closeQueue(channel: string): Observable<ChannelDto> {
+    return this.httpClient.put<ChannelDto>(
+      `${this.apiBaseUrl}/api/channels/${channel}`,
+      {
+        queueOpen: false,
+      },
+      {
+        withCredentials: true,
+      },
+    );
+  }
+
+  openQueue(channel: string): Observable<ChannelDto> {
+    return this.httpClient.put<ChannelDto>(
+      `${this.apiBaseUrl}/api/channels/${channel}`,
+      {
+        queueOpen: true,
+      },
+      {
+        withCredentials: true,
+      },
+    );
+  }
+
+  getGames(): Observable<GameDto[]> {
+    return this.httpClient.get<GameDto[]>(`${this.apiBaseUrl}/api/games`);
+  }
+
+  setGame(channel: string, gameId: number) {
+    return this.httpClient.put<ChannelDto>(
+      `${this.apiBaseUrl}/api/channels/${channel}`,
+      {
+        game: { id: gameId },
+      },
       {
         withCredentials: true,
       },
