@@ -1,4 +1,4 @@
-import { Controller, Get, Res, Inject, Query } from '@nestjs/common';
+import { Controller, Get, Res, Inject, Query, Logger } from '@nestjs/common';
 import { AuthService } from '../services/auth.service';
 import { Response } from 'express';
 
@@ -9,7 +9,6 @@ import {
   TWITCH_CLIENT_ID,
   TWITCH_CLIENT_SECRET,
 } from '../../../injection-tokens';
-import * as crypto from 'crypto';
 import { HttpService } from '@nestjs/axios';
 import { firstValueFrom } from 'rxjs';
 import { UserService } from '../../data-store/services/user.service';
@@ -17,6 +16,8 @@ import { AuthSource } from '../../data-store/entities/user-auth-source.entity';
 
 @Controller('auth/twitch')
 export class TwitchAuthController {
+  private logger: Logger = new Logger(this.constructor.name);
+
   constructor(
     private authService: AuthService,
     @Inject(JWT_COOKIE_NAME) private jwtCookieName: string,
@@ -29,10 +30,16 @@ export class TwitchAuthController {
   ) {}
 
   @Get('/')
-  authTwitchStart(@Res() res: Response) {
-    // FIXME: Make sure to store this somewhere to verify it when it returns.\
-    const state = crypto.randomBytes(20).toString('hex');
-    const url = `https://id.twitch.tv/oauth2/authorize?response_type=code&client_id=${this.twitchClientId}&redirect_uri=${this.baseUrl}/auth/twitch/return&state=${state}`;
+  authTwitchStart(@Query('mode') mode: string, @Res() res: Response) {
+    const stateKeyValues = new Map<string, string>();
+    if (mode) {
+      stateKeyValues.set('mode', mode);
+    }
+    const stateValue = this.authService.generateStateValue();
+
+    this.authService.storeState(stateValue, stateKeyValues);
+
+    const url = `https://id.twitch.tv/oauth2/authorize?response_type=code&client_id=${this.twitchClientId}&redirect_uri=${this.baseUrl}/auth/twitch/return&state=${stateValue}`;
 
     res.redirect(url);
   }
@@ -50,6 +57,17 @@ export class TwitchAuthController {
     // - Code exchange failed
     // - API call to get user information failed.
 
+    const stateKeyValues = this.authService.getState(state);
+    if (!stateKeyValues) {
+      // Invalid state - run them through authentication again.
+      this.logger.debug(
+        'Twitch auth failed: State value was not found in stateMap',
+        { state: state, code: code, scope: scope },
+      );
+      response.redirect('/auth/twitch');
+      return;
+    }
+
     // Send server-server request to get actual oauth token.
     const oauthCodeResponse = await firstValueFrom(
       this.httpService.post('https://id.twitch.tv/oauth2/token', {
@@ -60,6 +78,7 @@ export class TwitchAuthController {
         redirect_uri: `${this.baseUrl}/auth/twitch/return`,
       }),
     );
+    // FIXME: Add error handling if request failed, or code was invalid
 
     const tokenData = oauthCodeResponse.data;
 
@@ -72,6 +91,7 @@ export class TwitchAuthController {
         },
       }),
     );
+    // FIXME: Add error handling if API request failed.
 
     const user = await this.userService.loginUser(
       AuthSource.TWITCH,
@@ -86,18 +106,18 @@ export class TwitchAuthController {
 
     response.cookie(this.jwtCookieName, this.authService.getJwt(user));
 
-    // We'll need to detect what context we're working with, whether we're going back to electron, or going to launch this
-    // thing in the browser.
+    if (stateKeyValues.has('mode')) {
+      switch (stateKeyValues.get('mode')) {
+        case 'authcode':
+          // Return an auth code in the result that any client can use to get
+          // a JWT.
+          response.send('Paste this code into the client app: ' + authCode);
+          return;
+      }
+    }
 
-    // For now, we're just spitting back one of these auth codes for the electron app to use.
-    // FIXME: Replace this with a proper redirect when I work that out.
-
-    //response.send(authCode);
-
-    // Do electron redirect here.
+    // Assume it's an electron app and we redirect to the app url
     response.redirect(`requestobot://?authCode=${authCode}`);
-    // response.redirect('/launch');
+    return;
   }
 }
-
-// requestobot:///?authCode=dfd8deeb2bd471c7cc177cc40b19242b
