@@ -1,5 +1,5 @@
-import {ChatMessage} from "../services/chat-message";
+import { ChatMessage } from '../services/chat-message';
 
 export class ChatMessageReceiveEvent {
-    chatMessage: ChatMessage;
+  chatMessage: ChatMessage;
 }
